Refactored services user group, user and vehicle to work with fetch wrapper

+++ TravelBuddies/UserGroup/Commands/DeleteUserGroup/DeleteUserGroupHandler.cs
@@ -8,6 +8,7 @@
     using TravelBuddies.Application.Common.Interfaces.Repository;
 	using TravelBuddies.Application.Common.Exceptions.NotFound;
 	using TravelBuddies.Application.Common.Exceptions.Forbidden;
+	using Microsoft.EntityFrameworkCore;
 
 	public class DeleteUserGroupHandler : BaseHandler, IRequestHandler<DeleteUserGroupCommand, Task>
 	{
@@ -28,6 +29,24 @@
 			{
 				throw new ApplicationUserNotInGroupException(
 					string.Format(ApplicationUserNotInGroupMessage, request.UserId, request.GroupId));
+			}
+
+			Group? group = await _repository
+				.All<Group>(g => g.Id == request.GroupId)
+				.Include(g => g.Post)
+				.Include(g => g.UsersGroups)
+				.FirstOrDefaultAsync();
+
+			if(group == null)
+			{
+				throw new GroupNotFoundException(
+					string.Format(GroupNotFoundMessage, request.GroupId));
+			}
+
+			if(group.UsersGroups.Count - 1 == 0)
+			{
+				group.IsDeleted = true;
+				group.Post.IsDeleted = true;
 			}
 
 			_repository.Delete(userGroup);
